feat: update pr in gitea (#530)
