+++ crates/release_plz_core/src/release_pr.rs
@@ -6,8 +6,7 @@ use git_cmd::Repo;
 use anyhow::{anyhow, Context};
 use tracing::{info, instrument};
 
-use crate::backend::GitClient;
-use crate::github_client::{contributors_from_commits, GitHubPr};
+use crate::backend::{contributors_from_commits, GitClient, GitPr};
 use crate::pr::{Pr, BRANCH_PREFIX};
 use crate::{
     copy_to_temp_dir, publishable_packages, update, GitBackend, UpdateRequest, UpdateResult,
@@ -38,7 +37,7 @@ pub async fn release_pr(input: &ReleasePrRequest) -> anyhow::Result<()> {
         .set_local_manifest(&local_manifest)
         .context("can't find temporary project")?;
     let (packages_to_update, _temp_repository) = update(&new_update_request)?;
-    let git_client = GitClient::new(&input.git)?;
+    let git_client = GitClient::new(input.git.clone())?;
     if !packages_to_update.is_empty() {
         let repo = Repo::new(new_manifest_dir)?;
         let there_are_commits_to_push = repo.is_clean().is_err();
@@ -51,80 +50,64 @@ pub async fn release_pr(input: &ReleasePrRequest) -> anyhow::Result<()> {
     Ok(())
 }
 
-async fn open_or_update_release_pr<'a>(
+async fn open_or_update_release_pr(
     local_manifest: &Path,
     packages_to_update: &[(Package, UpdateResult)],
-    git_client: &GitClient<'a>,
+    git_client: &GitClient,
     repo: &Repo,
 ) -> anyhow::Result<()> {
-    match &git_client {
-        GitClient::GitHub(gh_client) => {
-            let opened_release_prs = gh_client
-                .opened_prs(BRANCH_PREFIX)
+    let opened_release_prs = git_client
+        .opened_prs(BRANCH_PREFIX)
+        .await
+        .context("cannot get opened release-plz prs")?;
+    // Close all release-plz prs, except one.
+    let old_release_prs = opened_release_prs.iter().skip(1);
+    for pr in old_release_prs {
+        git_client
+            .close_pr(pr.number)
+            .await
+            .context("cannot close old release-plz prs")?;
+    }
+
+    match opened_release_prs.first() {
+        Some(pr) => {
+            let pr_commits = git_client
+                .pr_commits(pr.number)
                 .await
-                .context("cannot get opened release-plz prs")?;
-            // Close all release-plz prs, except one.
-            let old_release_prs = opened_release_prs.iter().skip(1);
-            for pr in old_release_prs {
-                gh_client
+                .context("cannot get commits of release-plz pr")?;
+            let pr_contributors = contributors_from_commits(&pr_commits);
+            if pr_contributors.is_empty() {
+                // There are no contributors, so we can force-push
+                // in this PR, because we don't care about the git history.
+                let update_outcome = update_pr(pr, pr_commits.len(), repo);
+                if let Err(e) = update_outcome {
+                    tracing::error!("cannot update release pr {}: {}. I'm closing the old release pr and opening a new one", pr.number, e);
+                    git_client
+                        .close_pr(pr.number)
+                        .await
+                        .context("cannot close old release-plz prs")?;
+                    create_pr(git_client, repo, packages_to_update, local_manifest).await?
+                }
+            } else {
+                // There's a contributor, so we don't want to force-push in this PR.
+                // We close it because we want to save the contributor's work.
+                // TODO improvement: check how many lines the commit added, if no lines (for example a merge to update the branch),
+                //      then don't count it as a contributor.
+                info!("closing pr {} to preserve git history", pr.html_url);
+                git_client
                     .close_pr(pr.number)
                     .await
                     .context("cannot close old release-plz prs")?;
-            }
-
-            match opened_release_prs.first() {
-                Some(pr) => {
-                    let pr_commits = gh_client
-                        .pr_commits(pr.number)
-                        .await
-                        .context("cannot get commits of release-plz pr")?;
-                    let pr_contributors = contributors_from_commits(&pr_commits);
-                    if pr_contributors.is_empty() {
-                        // There are no contributors, so we can force-push
-                        // in this PR, because we don't care about the git history.
-                        let update_outcome = update_pr(pr, pr_commits.len(), repo);
-                        if let Err(e) = update_outcome {
-                            tracing::error!("cannot update release pr {}: {}. I'm closing the old release pr and opening a new one", pr.number, e);
-                            gh_client
-                                .close_pr(pr.number)
-                                .await
-                                .context("cannot close old release-plz prs")?;
-                            create_pr(git_client, repo, packages_to_update, local_manifest).await?
-                        }
-                    } else {
-                        // There's a contributor, so we don't want to force-push in this PR.
-                        // We close it because we want to save the contributor's work.
-                        // TODO improvement: check how many lines the commit added, if no lines (for example a merge to update the branch),
-                        //      then don't count it as a contributor.
-                        info!("closing pr {} to preserve git history", pr.html_url);
-                        gh_client
-                            .close_pr(pr.number)
-                            .await
-                            .context("cannot close old release-plz prs")?;
-                        create_pr(git_client, repo, packages_to_update, local_manifest).await?
-                    }
-                }
-                None => create_pr(git_client, repo, packages_to_update, local_manifest).await?,
+                create_pr(git_client, repo, packages_to_update, local_manifest).await?
             }
         }
-        GitClient::Gitea(_) => {
-            close_old_prs(git_client).await?;
-            create_pr(git_client, repo, packages_to_update, local_manifest).await?;
-        }
+        None => create_pr(git_client, repo, packages_to_update, local_manifest).await?,
     }
     Ok(())
 }
 
-async fn close_old_prs(git_client: &GitClient<'_>) -> anyhow::Result<()> {
-    git_client
-        .close_prs_on_branches(BRANCH_PREFIX)
-        .await
-        .context("cannot close old release-plz prs")?;
-    Ok(())
-}
-
 async fn create_pr(
-    git_client: &GitClient<'_>,
+    git_client: &GitClient,
     repo: &Repo,
     packages_to_update: &[(Package, UpdateResult)],
     local_manifest: &Path,
@@ -140,24 +123,24 @@ async fn create_pr(
     Ok(())
 }
 
-fn update_pr(pr: &GitHubPr, commits_number: usize, repository: &Repo) -> anyhow::Result<()> {
+fn update_pr(pr: &GitPr, commits_number: usize, repository: &Repo) -> anyhow::Result<()> {
     // save local work
     repository.git(&["stash", "--include-untracked"])?;
 
     reset_branch(pr, commits_number, repository).map_err(|e| {
         // restore local work
-        if let Err(e) = repository.checkout_stash() {
+        if let Err(e) = repository.stash_pop() {
             tracing::error!("cannot restore local work: {}", e);
         }
         e
     })?;
-    repository.checkout_stash()?;
+    repository.stash_pop()?;
     force_push(pr, repository)?;
     info!("updated pr {}", pr.html_url);
     Ok(())
 }
 
-fn reset_branch(pr: &GitHubPr, commits_number: usize, repository: &Repo) -> anyhow::Result<()> {
+fn reset_branch(pr: &GitPr, commits_number: usize, repository: &Repo) -> anyhow::Result<()> {
     // sanity check to avoid doing bad things on non-release-plz branches
     anyhow::ensure!(pr.branch().starts_with(BRANCH_PREFIX), "wrong branch name");
 
@@ -177,7 +160,7 @@ fn reset_branch(pr: &GitHubPr, commits_number: usize, repository: &Repo) -> anyh
     Ok(())
 }
 
-fn force_push(pr: &GitHubPr, repository: &Repo) -> anyhow::Result<()> {
+fn force_push(pr: &GitPr, repository: &Repo) -> anyhow::Result<()> {
     let changes_expect_typechanges = repository.changes_except_typechanges()?;
     repository.add(&changes_expect_typechanges)?;
     repository.commit("chore: release")?;
