+++ crates/release_plz_core/src/repo_url.rs
@@ -54,10 +54,11 @@ impl RepoUrl {
     }
 
     pub fn gitea_api_url(&self) -> String {
+        let v1 = "api/v1/";
         if let Some(port) = self.port {
-            format!("{}://{}:{}/api/v1", self.scheme, self.host, port)
+            format!("{}://{}:{}/{v1}", self.scheme, self.host, port)
         } else {
-            format!("{}://{}/api/v1", self.scheme, self.host)
+            format!("{}://{}/{v1}", self.scheme, self.host)
         }
     }
 }
